+++ pkg/api/types.go
@@ -115,7 +115,7 @@ type RouterProfile struct {
 // AgentPoolProfile represents configuration of OpenShift cluster VMs.
 type AgentPoolProfile struct {
 	Name       string `json:"name,omitempty"`
-	Count      int    `json:"count,omitempty"`
+	Count      int64  `json:"count,omitempty"`
 	VMSize     VMSize `json:"vmSize,omitempty"`
 	SubnetCIDR string `json:"subnetCidr,omitempty"`
 	OSType     OSType `json:"osType,omitempty"`
